+++ src/src/MatrixAPIInterface.java
@@ -3,9 +3,9 @@ package src;
 import java.util.List;
 
 public interface MatrixAPIInterface{
-  int[][] generateMatrix(int rows, int cols);
+  long[][] generateMatrix(int rows, int cols);
 
-  int[][] multiplyMatrices(int[][] matrix1, int[][] matrix2);
+  long[][] multiplyMatrices(long[][] matrix1, long[][] matrix2);
 
-List<int[][]> multiplyMatrices(List<int[][]> generatedMatrices);
+List<long[][]> multiplyMatrices(List<long[][]> generatedMatrices);
 }
