+++ test/tests/TestMatrixExceptionHandling.java
@@ -12,8 +12,8 @@ public class TestMatrixExceptionHandling {
   public void testMultiplyMatrices_ThrowsExceptionForIncompatibleDimensions() {
     MatrixImplementation matrixImplementation = new MatrixImplementation();
     
-    int[][] matrix1 = {{1, 2}, {3, 4}}; 
-    int[][] matrix2 = {{1, 2, 3}};  
+    long[][] matrix1 = {{1, 2}, {3, 4}}; 
+    long[][] matrix2 = {{1, 2, 3}};  
 
 	// Expect an IllegalArgumentException due to incompatible dimensions
     Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () -> {
@@ -26,7 +26,7 @@ public class TestMatrixExceptionHandling {
   public void testPrintMatrix_ThrowsExceptionForMismatchedDimensions() {
     MatrixImplementation matrixImplementation = new MatrixImplementation();
 		    
-    int[][] matrix = {{1, 2}, {3, 4}}; // 2x2 matrix
+    long[][] matrix = {{1, 2}, {3, 4}}; // 2x2 matrix
     // Expect an IllegalArgumentException due to mismatched dimensions (3x3 requested but matrix is 2x2)
     Exception exception = Assertions.assertThrows(IllegalArgumentException.class, () -> {
       matrixImplementation.printMatrix(matrix, 3, 3);
